[refactor] `FeatureConventionPlugin`
